requirements.txt
torch
torchvision
pandas
numpy
Pillow
pillow-heif
scikit-learn
